Make mdapi conversion the default mode
Retrieving a changeset downloads all the components along with it. Converting the metadata that we already have will save us having to kick off another retrieve from salesforce, and hopefully save a bit of time.

The source mode will stay around until we need it, or prove that we don't need it.